CMake: Change (implicit) NOTICE-mode messages to STATUS
Many messages not requiring user attention are simply written as

	message("text")

which is an implicit

	message(NOTICE "text")

with some implications, for example, they are written to stderr
when using pure cmake.
Moreover, ccmake users have to explicitly acknowledge these
messages, which is cumbersome to use.

This commit changes these messages to mode STATUS.

While at it, slightly fix CMake coding style in locations
surrounding these changes.

Signed-off-by: Stephan Beyer <s-beyer@gmx.net>